Revert "Babel 7, add tree shaking, replace latest with env (#134)"
This reverts commit 443720b9e3dbb46799a17206c3edbe139fc724e9.
