+++ .gitignore
@@ -16,7 +16,6 @@ examples/*/yarn.lock
 .env.development.local
 .env.test.local
 .env.production.local
-.esm-cache
 
 npm-debug.log*
 yarn-debug.log*
