+++ src/commands/index.js
@@ -1,12 +1,62 @@
-/* eslint-disable */
-require = require('@std/esm')(module, {
-  esm: 'js',
-  cjs: true
-})
-require('@babel/register')({
-  babelrc: false,
-  presets: [require('../../babel-preset')],
-  cache: false
+import 'babel-register'
+import chalk from 'chalk'
+import fs from 'fs'
+
+const ignoredExtensions = [
+  'css',
+  'scss',
+  'styl',
+  'less',
+  'png',
+  'gif',
+  'jpg',
+  'jpeg',
+  'svg',
+  'woff',
+  'ttf',
+  'eot',
+  'otf',
+  'mp4',
+  'webm',
+  'ogg',
+  'mp3',
+  'wav',
+]
+ignoredExtensions.forEach(ext => {
+  require.extensions[`.${ext}`] = () => {}
 })
 
-module.exports.default = require('./main').default
+export default function () {
+  const cmd = process.argv[2]
+
+  if (['-v', '--version'].indexOf(cmd) !== -1) {
+    const packageJson = JSON.parse(fs.readFileSync(`${__dirname}/../../package.json`, 'utf8'))
+    return console.log(packageJson.version)
+  }
+
+  if (cmd === 'start') {
+    return require('./start').default()
+  }
+
+  if (cmd === 'build') {
+    return require('./build').default()
+  }
+
+  if (cmd === 'create') {
+    return require('./create').default(process.argv[3])
+  }
+
+  console.log(
+    `
+Usage: react-static <command>
+
+- ${chalk.green('create')}  -  create a new project
+- ${chalk.green('start')}  -  start the development server
+- ${chalk.green('build')}  -  build site for production
+
+Options:
+
+    -v, --version output the version number
+`,
+  )
+}
